+++ src/BackupTrait.php
@@ -54,9 +54,9 @@ trait BackupTrait
      */
     public static function getCompression(string $path): ?string
     {
-        $extension = self::getExtension($path);
+        $Compression = Compression::tryFromFilename($path);
 
-        return self::getValidCompressions()[$extension] ?? null;
+        return $Compression && $Compression !== Compression::None ? lcfirst($Compression->name) : null;
     }
 
     /**
@@ -93,15 +93,7 @@ trait BackupTrait
      */
     public static function getExtension(string $path): ?string
     {
-        $path = strtolower($path);
-
-        foreach (array_keys(DATABASE_BACKUP_EXTENSIONS) as $extension) {
-            if (str_ends_with($path, '.' . $extension)) {
-                return $extension;
-            }
-        }
-
-        return null;
+        return Compression::tryFromFilename($path)?->value;
     }
 
     /**
@@ -109,9 +101,22 @@ trait BackupTrait
      *
      * @return array<string, string> An array with extensions as keys and compressions as values
      * @since 2.4.0
+     * @deprecated 2.13.5 the `BackupTrait::getValidCompressions()` method is deprecated. Will be removed in a future release
      */
     public static function getValidCompressions(): array
     {
-        return array_filter(DATABASE_BACKUP_EXTENSIONS);
+        deprecationWarning(
+            '2.13.5',
+            'The `BackupTrait::getValidCompressions()` method is deprecated. Will be removed in a future release'
+        );
+
+        return array_map(callback: 'lcfirst', array: array_column(
+            array: array_filter(
+                array: Compression::cases(),
+                callback: fn (Compression $Compression): bool => $Compression != Compression::None,
+            ),
+            column_key: 'name',
+            index_key: 'value'
+        ));
     }
 }