+++ tests/TestCase/Utility/BackupExportTest.php
@@ -17,11 +17,17 @@ namespace DatabaseBackup\Test\TestCase\Utility;
 
 use Cake\Core\Configure;
 use Cake\Event\EventList;
+use DatabaseBackup\Compression;
 use DatabaseBackup\Driver\Sqlite;
 use DatabaseBackup\TestSuite\TestCase;
 use DatabaseBackup\Utility\BackupExport;
+use InvalidArgumentException;
+use PHPUnit\Framework\Attributes\Test;
+use PHPUnit\Framework\Attributes\TestWith;
+use PHPUnit\Framework\Attributes\WithoutErrorHandler;
 use Symfony\Component\Process\Exception\ProcessTimedOutException;
 use Symfony\Component\Process\Process;
+use ValueError;
 
 /**
  * BackupExportTest class.
@@ -47,19 +53,48 @@ class BackupExportTest extends TestCase
     }
 
     /**
-     * Test for `compression()` method. This also tests for `$extension` property.
-     *
-     * @test
      * @uses \DatabaseBackup\Utility\BackupExport::compression()
      */
-    public function testCompression(): void
+    #[Test]
+    #[TestWith([Compression::None])]
+    #[TestWith([Compression::Gzip])]
+    #[TestWith([Compression::Bzip2])]
+    public function testCompression(Compression $Compression): void
     {
-        $this->BackupExport->compression('bzip2');
-        $this->assertSame('bzip2', $this->BackupExport->compression);
-        $this->assertSame('sql.bz2', $this->BackupExport->extension);
+        $this->BackupExport->compression($Compression);
+        $this->assertSame($Compression, $this->BackupExport->compression);
+    }
 
-        //With an invalid type
-        $this->expectExceptionMessage('Invalid compression type');
+    #[Test]
+    #[TestWith([Compression::None, null])]
+    #[TestWith([Compression::Gzip, 'gzip'])]
+    #[TestWith([Compression::Bzip2, 'bzip2'])]
+    public function testCompressionAsStringOrNull(Compression $ExpectedCompression, ?string $compressionAsStringOrNull): void
+    {
+        $this->BackupExport->compression($compressionAsStringOrNull);
+        $this->assertSame($ExpectedCompression, $this->BackupExport->compression);
+    }
+
+    #[Test]
+    #[TestWith([null])]
+    #[TestWith(['gzip'])]
+    #[TestWith(['bzip2'])]
+    #[WithoutErrorHandler]
+    public function testCompressionAsStringOrNullIsDeprecated(?string $compressionAsStringOrNull): void
+    {
+        $this->deprecated(function () use ($compressionAsStringOrNull): void {
+            $this->BackupExport->compression($compressionAsStringOrNull);
+        });
+    }
+
+    /**
+     * @uses \DatabaseBackup\Utility\BackupExport::compression()
+     */
+    #[Test]
+    public function testCompressionWithInvalidCompressionAsString(): void
+    {
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage('No valid `' . Compression::class . '` value was found starting from `invalidType`');
         $this->BackupExport->compression('invalidType');
     }
 
@@ -73,14 +108,12 @@ class BackupExportTest extends TestCase
     {
         $this->BackupExport->filename('backup.sql.bz2');
         $this->assertSame(Configure::read('DatabaseBackup.target') . 'backup.sql.bz2', $this->BackupExport->filename);
-        $this->assertSame('bzip2', $this->BackupExport->compression);
-        $this->assertSame('sql.bz2', $this->BackupExport->extension);
+        $this->assertSame(Compression::Bzip2, $this->BackupExport->compression);
 
         //Compression is ignored, because there's a filename
         $this->BackupExport->compression('gzip')->filename('backup.sql.bz2');
         $this->assertSame('backup.sql.bz2', basename($this->BackupExport->filename));
-        $this->assertSame('bzip2', $this->BackupExport->compression);
-        $this->assertSame('sql.bz2', $this->BackupExport->extension);
+        $this->assertSame(Compression::Bzip2, $this->BackupExport->compression);
 
         //Filename with `{$DATABASE}` pattern
         $this->BackupExport->filename('{$DATABASE}.sql');
@@ -99,8 +132,10 @@ class BackupExportTest extends TestCase
         $this->assertMatchesRegularExpression('/^\d{10}\.sql$/', basename($this->BackupExport->filename));
 
         //With invalid extension
-        $this->expectExceptionMessage('Invalid `txt` file extension');
-        $this->BackupExport->filename('backup.txt');
+        $filename = TMP . 'backup.txt';
+        $this->expectException(ValueError::class);
+        $this->expectExceptionMessage('No valid `' . Compression::class . '` value was found for filename `' . $filename . '`');
+        $this->BackupExport->filename($filename);
     }
 
     /**
@@ -180,10 +215,15 @@ class BackupExportTest extends TestCase
     public function testExportStoppedByBeforeExport(): void
     {
         $Driver = $this->createPartialMock(Sqlite::class, ['beforeExport']);
-        $Driver->method('beforeExport')->willReturn(false);
+        $Driver->method('beforeExport')
+            ->willReturn(false);
         $Driver->getEventManager()->on($Driver);
-        $BackupExport = $this->createPartialMock(BackupExport::class, ['getDriver']);
-        $BackupExport->method('getDriver')->willReturn($Driver);
+
+        $BackupExport = $this->getMockBuilder(BackupExport::class)
+            ->onlyMethods(['getDriver'])
+            ->getMock();
+        $BackupExport->method('getDriver')
+            ->willReturn($Driver);
         $this->assertFalse($BackupExport->export());
     }
 
@@ -192,7 +232,6 @@ class BackupExportTest extends TestCase
      *
      * @test
      * @throws \PHPUnit\Framework\MockObject\Exception
-     * @throws \ReflectionException
      * @uses \DatabaseBackup\Utility\BackupExport::export()
      */
     public function testExportOnFailure(): void
@@ -200,8 +239,11 @@ class BackupExportTest extends TestCase
         $expectedError = 'mysqldump: Got error: 1044: "Access denied for user \'root\'@\'localhost\' to database \'noExisting\'" when selecting the database';
         $this->expectExceptionMessage('Export failed with error message: `' . $expectedError . '`');
         $Process = $this->createConfiguredMock(Process::class, ['getErrorOutput' => $expectedError . PHP_EOL, 'isSuccessful' => false]);
-        $BackupExport = $this->createPartialMock(BackupExport::class, ['getProcess']);
-        $BackupExport->method('getProcess')->willReturn($Process);
+        $BackupExport = $this->getMockBuilder(BackupExport::class)
+            ->onlyMethods(['getProcess'])
+            ->getMock();
+        $BackupExport->method('getProcess')
+            ->willReturn($Process);
         $BackupExport->export();
     }
 
@@ -212,15 +254,17 @@ class BackupExportTest extends TestCase
      * @test
      * @uses \DatabaseBackup\Utility\BackupExport::export()
      * @throws \PHPUnit\Framework\MockObject\Exception
-     * @throws \ReflectionException
      */
     public function testExportExceedingTimeout(): void
     {
         $this->expectException(ProcessTimedOutException::class);
         $this->expectExceptionMessage('The process "dir" exceeded the timeout of 60 seconds');
         $ProcessTimedOutException = new ProcessTimedOutException(Process::fromShellCommandline('dir'), 1);
-        $BackupExport = $this->createPartialMock(BackupExport::class, ['getProcess']);
-        $BackupExport->method('getProcess')->willThrowException($ProcessTimedOutException);
+        $BackupExport = $this->getMockBuilder(BackupExport::class)
+            ->onlyMethods(['getProcess'])
+            ->getMock();
+        $BackupExport->method('getProcess')
+            ->willThrowException($ProcessTimedOutException);
         $BackupExport->export();
     }
 }